requirements.txt
fire
flask
pillow
